extract getSuspendedSales
