+++ monolith/interactors/getSuspendedSales.ts
@@ -1,0 +1,45 @@
+import { MongoClient } from 'mongodb';
+import { ClubhouseLocation } from './getJwt';
+import fetchDbName from '../lib/fetchDbName';
+import collectionFromLocation from '../lib/collectionFromLocation';
+const DATABASE_NAME = fetchDbName();
+
+const mongoOptions = {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+};
+
+/**
+ * Returns all suspended sales' ids, customer names, and notes (omitting card lists)
+ */
+async function getSuspendedSales(location: ClubhouseLocation) {
+    const client = await new MongoClient(process.env.MONGO_URI, mongoOptions);
+
+    try {
+        await client.connect();
+
+        const db = client
+            .db(DATABASE_NAME)
+            .collection(collectionFromLocation(location).suspendedSales);
+
+        const docs = await db.find(
+            {},
+            {
+                projection: {
+                    _id: 1,
+                    createdAt: 1,
+                    name: 1,
+                    notes: 1,
+                },
+            }
+        );
+
+        return await docs.toArray();
+    } catch (e) {
+        throw e;
+    } finally {
+        await client.close();
+    }
+}
+
+export default getSuspendedSales;
