+++ monolith/routes/auth.ts
@@ -17,6 +17,7 @@ import getSalesFromCardname from '../interactors/getSalesFromCardname';
 import getAllSales from '../interactors/getAllSales';
 import getFormatLegalities from '../interactors/getFormatLegalities';
 import addCardToInventoryReceiving from '../interactors/addCardToInventoryReceiving';
+import getSuspendedSales from '../interactors/getSuspendedSales';
 
 interface RequestWithUserInfo extends Request {
     locations: string[];
@@ -234,39 +235,6 @@ router.post('/receiveCards', async (req: RequestWithUserInfo, res) => {
         res.status(500).send(err);
     }
 });
-
-/**
- * Returns all suspended sales' ids, customer names, and notes (omitting card lists)
- */
-async function getSuspendedSales(location: ClubhouseLocation) {
-    const client = await new MongoClient(process.env.MONGO_URI, mongoOptions);
-
-    try {
-        await client.connect();
-
-        const db = client
-            .db(DATABASE_NAME)
-            .collection(collectionFromLocation(location).suspendedSales);
-
-        const docs = await db.find(
-            {},
-            {
-                projection: {
-                    _id: 1,
-                    createdAt: 1,
-                    name: 1,
-                    notes: 1,
-                },
-            }
-        );
-
-        return await docs.toArray();
-    } catch (e) {
-        throw e;
-    } finally {
-        await client.close();
-    }
-}
 
 /**
  * Retrieves one suspended sale
